final code

+++ app/models/users.py
@@ -1,11 +1,13 @@
 from sqlalchemy import Column, Integer, String, Boolean, DateTime, func
 from app.core.database import Base
+from sqlalchemy import BigInteger
 
 class User(Base):
     __tablename__ = "risk_dashboard_users"
     __table_args__ = {"schema": "rt"}  # Crucial: Point to 'rt' schema
 
-    id = Column(Integer, primary_key=True, index=True)
+    # id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     username = Column(String, unique=True, index=True)
     email = Column(String, unique=True, index=True)
     password_hash = Column(String)
